Fix SfChronicle (reload loop)
Exclusion from localstorage.clear (contentScript.js).

+++ contentScript.js
@@ -1,4 +1,11 @@
-window.localStorage.clear();
+var arr_localstorage_hold = ['sfchronicle.com'];
+var localstorage_hold = arr_localstorage_hold.some(function(url) {
+	return window.location.href.indexOf(url) !== -1;
+});
+
+if (!localstorage_hold){
+	window.localStorage.clear();
+}
 
 if (location.hostname.endsWith('rep.repubblica.it')) {
     if (location.href.includes('/pwa/')) {
@@ -168,11 +175,6 @@ if (window.location.href.indexOf("canberratimes.com.au") !== -1) {
 if (window.location.href.indexOf("ledevoir.com") !== -1) {
         const counter = document.querySelector('.full.hidden-print.popup-msg');
         removeDOMElement(counter);
-}
-
-if (window.location.href.includes('ft.com')) {
-    const cookie_banner = document.querySelector('.n-messaging-banner__outer');
-    removeDOMElement(cookie_banner);
 }
 
 if (window.location.href.indexOf("thehindu.com") !== -1) {
